+++ cmd/main.go
@@ -12,7 +12,7 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-const version string = "1.1.1"
+const version string = "1.1.2"
 
 func main() {
 	log.SetOutput(os.Stdout)