+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [v1.1.2](https://github.com/containeroo/SyncFlaer/tree/v1.1.2) (2021-02-02)
+
+[All Commits](https://github.com/containeroo/SyncFlaer/compare/v1.1.1...v1.1.2)
+
+**Bug fixes:**
+
+- fix an issue where changing a `cloudflare.default` config resulted in an unexpected error
+- various bug fixes for delete grace feature
+
 ## [v1.1.1](https://github.com/containeroo/SyncFlaer/tree/v1.1.1) (2021-02-01)
 
 [All Commits](https://github.com/containeroo/SyncFlaer/compare/v1.1.0...v1.1.1)